Do not use package object for syntaxes
