+++ cats/src/main/scala/magnolify/cats/GroupDerivation.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package magnolify.cats.semiauto
+package magnolify.cats
 
 import cats.Group
 import magnolia1._
