+++ cats/src/main/scala/magnolify/cats/HashDerivation.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package magnolify.cats.semiauto
+package magnolify.cats
 
 import cats.Hash
 import magnolia1._
