+++ cats/src/main/scala/magnolify/cats/ShowDerivation.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package magnolify.cats.semiauto
+package magnolify.cats
 
 import cats.Show
 import magnolia1._
