+++ guava/src/main/scala/magnolify/guava/FunnelDerivation.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package magnolify.guava.semiauto
+package magnolify.guava
 
 import com.google.common.base.Charsets
-import com.google.common.hash.{Funnel, Funnels, PrimitiveSink}
+import com.google.common.hash.{Funnel, PrimitiveSink}
 import magnolia1._
 
 import scala.annotation.nowarn
@@ -51,36 +51,4 @@ object FunnelDerivation {
   def by[T, S](f: T => S)(implicit fnl: Funnel[S]): Funnel[T] = new Funnel[T] {
     override def funnel(from: T, into: PrimitiveSink): Unit = fnl.funnel(f(from), into)
   }
-}
-
-trait FunnelImplicits {
-  private def funnel[T](f: (PrimitiveSink, T) => Unit): Funnel[T] = new Funnel[T] {
-    override def funnel(from: T, into: PrimitiveSink): Unit = f(into, from)
-  }
-
-  implicit val intFunnel: Funnel[Int] = Funnels.integerFunnel().asInstanceOf[Funnel[Int]]
-  implicit val longFunnel: Funnel[Long] = Funnels.longFunnel().asInstanceOf[Funnel[Long]]
-  implicit val bytesFunnel: Funnel[Array[Byte]] = Funnels.byteArrayFunnel()
-  implicit val booleanFunnel: Funnel[Boolean] = funnel[Boolean](_.putBoolean(_): @nowarn)
-  implicit val byteFunnel: Funnel[Byte] = funnel[Byte](_.putByte(_): @nowarn)
-  implicit val charFunnel: Funnel[Char] = funnel[Char](_.putChar(_): @nowarn)
-  implicit val shortFunnel: Funnel[Short] = funnel[Short](_.putShort(_): @nowarn)
-
-  implicit def charSequenceFunnel[T <: CharSequence]: Funnel[T] =
-    Funnels.unencodedCharsFunnel().asInstanceOf[Funnel[T]]
-
-  // There is an implicit Option[T] => Iterable[T]
-  implicit def iterableFunnel[T, C[_]](implicit
-    fnl: Funnel[T],
-    ti: C[T] => Iterable[T]
-  ): Funnel[C[T]] =
-    funnel { (sink, from) =>
-      var i = 0
-      from.foreach { x =>
-        fnl.funnel(x, sink)
-        i += 1
-      }
-      // inject size to distinguish `None`, `Some("")`, and `List("", "", ...)`
-      sink.putInt(i): @nowarn
-    }
 }
